### 2020-02-01 #2057 @arfon
@pstrk - many thanks for your submission to JOSS. Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
