### 2020-02-03 #2057 @diversemix
How does this work fit into to other's work in this field?

For example `pyunicorn`  [Unified functional network and nonlinear time series analysis for complex systems science: The pyunicorn package](https://aip.scitation.org/doi/10.1063/1.4934554)
