### 2020-02-04 #2057 @pstrk
> How does this work fit into to other's work in this field?
>
> For example `pyunicorn` [Unified functional network and nonlinear time series analysis for complex systems science: The pyunicorn package](https://aip.scitation.org/doi/10.1063/1.4934554)

I was not familiar with pyunicorn before and did just quickly through the paper. I think there is some confusion about the term "complex network".  According to e.g. wikipedia complex networks are graphs with non-trivial topological features.  In our case however we model networks where the edges have complex valued weights (i.e. phase and amplitude) and are delayed (meaning it takes a number of timesteps until a signal propagtes through the edge). Our software package is focused on returning all paths through the network, leading from a set of "input nodes" to a set of "output nodes", which are both defined by the user. If there are recurrent paths (i.e. loops) in the network, our software will return the paths until the amplitude decays below a user provided threshold (loops must have gain < 1, otherwise the analysis will not converge).  We do not provide any network statistics/graph measures.

Pyunicorn, iGraph etc. are very general packages to model complex networks resp. graphs (not sure if they do support complex valued edges though), while our package is focused on a very specific type of network/graph (complex valued and delayed edges, linear nodes) and is focused on modelling the signal propagation through these kind of networks. Additionally we provide a framework that supports the use of symbolic numbers for all edge properties, that can then be set to a target value using a feed dictionary, when analyzing the network or when calculating output signals.
