### 2020-02-05 5af75fa
typos in readme and contributing
