### 2020-02-05 0a37350
adds classifiers to the setup
