### 2020-02-05 d65a027
readme installation instruction updated with available pip package
