+++ READMEpypi.md
@@ -15,7 +15,7 @@ COLNA can inject complex valued signals to the network and the compute the resul
 
 The core functionality of COLNA is visualized in the figure below.
 
-![](docs/_images/colna_features_extended_plain.svg)
+.. image::  https://github.com/IBM/complex-linear-network-analyzer/blob/master/docs/_images/colna_features_extended_plain.svg
 
 Documentation
 -------------
@@ -43,13 +43,3 @@ Requirements
 Numpy, Scipy, Matplotlib, tqdm
 
 For visualization: Graphviz + Graphviz Python Package (see installation instructions for details)
-
-Contributing
-------------
-
-Contributions are very welcome! Please follow the the [contribution guidelines](CONTRIBUTING.md).
-
-Authors and Maintainers
------------------------
-* Pascal Stark
-* Lorenz Mueller
