### 2020-02-05 #2073 @whedon
[REVIEW]: COLNA: A Python Package to Analyze Complex, Linear Networks
**Submitting author:** @pstrk (<a href="http://orcid.org/0000-0002-7389-5592">Pascal Stark</a>)
**Repository:** <a href="https://github.com/IBM/complex-linear-network-analyzer" target ="_blank">https://github.com/IBM/complex-linear-network-analyzer</a>
**Version:** v1.0
