### 2020-02-08 #2073 @brandondube
Hello,

I am working the review - so far so good.  Some early comments:

- Lorenz, the second author, does not show up as a contributor on the GitHub repository.  Is this an error?  Or are their contributions made in a different way?

- the paper includes a nice statement of need, but I did not find a "what can I do with this" (application level, not technical) in the first few pages of the docs.  Perhaps this could be added, something like "these features can be used to...", where "to" isn't analyze a network, but, say, design a PIC.

- I see there are tests, but there is no coverage report.  I ran pytest with pytest-cov locally, and got 73% coverage.  Perhaps a statement of test coverage could be added to the repository, or a coveralls badge if the authors are interested in that sort of bling.

- I tried running a few examples.  There are various errors, some "cosmetic" in nature, others seem to disrupt the core functionality.  -- see [here](https://gist.github.com/brandondube/417b0c808521e94be256a23d389038f5).  For reproducibility, numpy 1.18.1, scipy 1.3.2, win10/amd64, numpy and scipy from conda.

- continuing with examples, the "small recurrent network" one seems to expose a performance issue, or just a lack of understanding on my part.

- - It begins at about 4,000 paths per second, and now 4 minutes later it is at about 100 paths per second and is using a hefty 18GB of RAM.  Is this expected?  Perhaps add a note in the examples that this one is particularly resource intensive.  I stopped it at `Network evaluation in progress: 140185paths [07:24, 78.09paths/s]` to try more examples.

- - I'm running this on a fairly beefy computer (i7-9700k, 32GB of RAM) and also see it is single threaded.  Is threading deferred to the user (for example, `OMP_NUM_THREADS` and related environment variables), or is there a specific enable/disable configuration for the package?
