### 2020-02-08 #2073 @pstrk
Thank you for your feedback, I will quickly answer the non-technical questions.

> Lorenz, the second author, does not show up as a contributor on the GitHub repository. Is this an error? Or are their contributions made in a different way?

Lorenz contributed about half of the code. We developed the code on a private Github instance, before moving it to the public Github. We did not want to include the full development history, as it would include some files that we can not yet share. For simplicity I just duplicated the code locally, removed all these files and uploaded the new version using my Github account. This is the reason why Lorenz does not show up as a contributer.

> the paper includes a nice statement of need, but I did not find a "what can I do with this" (application level, not technical) in the first few pages of the docs. Perhaps this could be added, something like "these features can be used to...", where "to" isn't analyze a network, but, say, design a PIC.

Thank you for the comment, we will add a statment of need in the documentation as well.

> I see there are tests, but there is no coverage report. I ran pytest with pytest-cov locally, and got 73% coverage. Perhaps a statement of test coverage could be added to the repository, or a coveralls badge if the authors are interested in that sort of bling.

I'll work on that as well. All visualization features are not tested, we will think about a way to verify them.

> I tried running a few examples. There are various errors, some "cosmetic" in nature, others seem to disrupt the core functionality. -- see here. For reproducibility, numpy 1.18.1, scipy 1.3.2, win10/amd64, numpy and scipy from conda.

I suggest we move this to an issue discussion in our repository. I will try to reconstruct the errors and check all the examples.

> continuing with examples, the "small recurrent network" one seems to expose a performance issue, or just a lack of understanding on my part.

I just double checked the example in the docs and there was a small typo: Edge('a', 'b', phase=0.5, attenuation=1.95, delay=1.0) should be Edge('a', 'b', phase=0.5, attenuation=0.95, delay=1.0). This leads to gain in the loop and will not converge. Eventually the evaluation would stop (after going thorugh 1e6 paths), but this will take a long time and way to much memory in the case of such a small network. I will fix the documentation accordingly.

The network evaluation is so far indeed single threaded as you will run it only once in our typical use case, so we can afford the single thread performance limitation. Testbenchs (used to compute output signals) do support multithreading, as we often run them multiple times with different feed dictionaries.
