### 2020-02-08 a1f8b85
fixes type conversion, prepare for release v1.0.2
