### 2020-02-08 IBM/complex-linear-network-analyzer#13 @pstrk
Create output directories if they do not exist
Output directories (for visualizations and pretty print export) shoud be created automatically, if they do not yet exist. Using pathlib syntax this can be easily achived in a single line.

from pathlib import Path
Path("/test/output").mkdir(parents=True, exist_ok=True)

This issue was mentioned first by @brandondube in the JOSS review thread (https://github.com/openjournals/joss-reviews/issues/2073#issuecomment-583763095)
