### 2020-02-09 #2073 @brandondube
👍 for the nontechnical topics.  For the gain issue -- perhaps you could add a 'feasibility' test that can ensure a network is not divergent (within reason).
