### 2020-02-11 9998886
examples removed figure output
Example scripts should run without any setup.
