+++ examples/symbolicfeedforwardwithtestbench.py
@@ -79,5 +79,5 @@ plt.xlabel('Time')
 plt.ylabel('|x|')
 plt.legend(['Input', 'Output C', 'Output D', 'Output C (Feed Dict 2)', 'Output D (Feed Dict 2)'], loc='lower left')
 plt.grid()
-plt.savefig('./visualizations/symnum_feedforward_tb_output.svg')
+# plt.savefig('./visualizations/symnum_feedforward_tb_output.svg')
 plt.show()
