+++ examples/feedforwardwithtestbench.py
@@ -72,5 +72,5 @@ plt.xlabel('Time')
 plt.ylabel('|x|')
 plt.legend(['Input', 'Output C', 'Output D'], loc='lower left')
 plt.grid()
-plt.savefig('./visualizations/feedforward_with_testbench_output.svg')
+# plt.savefig('./visualizations/feedforward_with_testbench_output.svg')
 plt.show()
