### 2020-02-11 b962c46
automatic output path generation
updated documentation
