### 2020-02-11 #2073 @pstrk
> I tried running a few examples. There are various errors, some "cosmetic" in nature, others seem to disrupt the core functionality. -- see here. For reproducibility, numpy 1.18.1, scipy 1.3.2, win10/amd64, numpy and scipy from conda.

I uploaded a new version (v1.0.2) that should solve the issues you observed. Please let me know in case it does not work.
