### 2020-02-20 #2073 @DiegoJRR
Hi, I am almost done with the review.

I did not find real world examples in the documentation, or some direct application for the software. This would add a lot to its usability.

 Apart from this, everything LGTM! I had no problems running the examples and tests. Excellent work
