### 2020-02-24 e70714f
Added method to get evaluated network result and reduced result
including documentation

Signed-Off: Pascal Stark <crk@zurich.ibm.com>
