+++ designexamples/mziweight.py
@@ -1,0 +1,125 @@
+# +-----------------------------------------------------------------------------+
+# |  Copyright 2019-2020 IBM Corp. All Rights Reserved.                         |
+# |                                                                             |
+# |  Licensed under the Apache License, Version 2.0 (the "License");            |
+# |  you may not use this file except in compliance with the License.           |
+# |  You may obtain a copy of the License at                                    |
+# |                                                                             |
+# |      http://www.apache.org/licenses/LICENSE-2.0                             |
+# |                                                                             |
+# |  Unless required by applicable law or agreed to in writing, software        |
+# |  distributed under the License is distributed on an "AS IS" BASIS,          |
+# |  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   |
+# |  See the License for the specific language governing permissions and        |
+# |  limitations under the License.                                             |
+# +-----------------------------------------------------------------------------+
+# |  Authors: Lorenz K. Mueller, Pascal Stark                                   |
+# +-----------------------------------------------------------------------------+
+"""
+Models a Mach Zehnder Modulator.
+"""
+
+from colna.analyticnetwork import Network, Edge, SymNum
+from colna.analyticnetwork import Testbench
+from colna.analyticnetwork import PhysicalNetwork, Device, DeviceLink
+from colna.lossconversion import dBcm_to_attenuation
+import numpy as np
+import matplotlib.pyplot as plt
+
+# Global defaults
+wavelength = 1.55e-6
+neff = 2.6
+
+# Splitter configuration
+splitter_length = 10e-6
+splitter_phase_default = 2 * np.pi / wavelength * neff * splitter_length
+splitter_efficiency00 = SymNum(name='split_{00}', default=1 / np.sqrt(2), product=True)
+splitter_efficiency01 = SymNum(name='split_{01}', default=1 / np.sqrt(2), product=True)
+splitter_phase00 = SymNum(name='split_{\phi00}', default=splitter_phase_default, product=False)
+splitter_phase01 = SymNum(name='split_{\phi01}', default=splitter_phase_default, product=False)
+
+splitter = Device('splitter')
+splitter.init_from_phase_and_attenuation_matrix(np.array([[splitter_efficiency00, splitter_efficiency01]]),
+                                                np.array([[splitter_phase00, splitter_phase01]]), delay=0)
+splitter.add_input('i0')
+
+# Combiner configuration
+combiner_length = 10e-6
+combiner_phase_default = 2 * np.pi / wavelength * neff * combiner_length
+combiner_efficiency00 = SymNum(name='comb_{00}', default=1 / np.sqrt(2), product=True)
+combiner_efficiency01 = SymNum(name='comb_{10}', default=1 / np.sqrt(2), product=True)
+combiner_phase00 = SymNum(name='comb_{\phi00}', default=combiner_phase_default, product=False)
+combiner_phase10 = SymNum(name='comb_{\phi10}', default=combiner_phase_default, product=False)
+
+combiner = Device('combiner')
+combiner.init_from_phase_and_attenuation_matrix(np.array([[combiner_efficiency00], [combiner_efficiency01]]),
+                                                np.array([[combiner_phase00],[combiner_phase10]]),
[…]
