### 2020-02-24 c6c5114
MZI Weight Design Example Documentation
Signed-Off: Pascal Stark <crk@zurich.ibm.com>
