### 2020-02-24 #2073 @pstrk
Hi,
Thank you for your feedback. I added a new design example (Mach Zehnder interferometer used as an optical weight), which should link more to the real world application scenarios of COLNA. Furthermore I improved the test coverage and added a few new convenience methods. I therefore released again a new version (1.0.3).

@arfon Do I need to do anything so that we can proceed with the publication?
