### 2020-02-25 #2073 @pstrk
Hi @arfon
Thank you for the information. I created a new release (V1.0.4) and the corresponding [Zenodo archive](https://doi.org/10.5281/zenodo.3686634). The DOI is 10.5281/zenodo.3686634.
